requirements.txt
owoify_py==2.0.0
